feat(id): 新增雪花分布式ID生成器，插入数据时系统自动生成ID和创建时间

+++ src/schema.rs
@@ -1,6 +1,6 @@
 table! {
     note_card (id) {
-        id -> Integer,
+        id -> BigInt,
         title -> Text,
         content -> Text,
         tip -> Text,
